refactor: extract shared UI helpers
